app/pyproject.toml
[project]
name = "swoop-intelligence"
version = "1.0.0"
description = "Chunk-based automotive service intelligence platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "supabase>=2.9.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.3.0",
    "ddgs>=0.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
